float版を実装

+++ 3_floating_point/main.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 
 int main() {
-    double x = 1.0;
+    float x = 1.0f;
     int count = 0;
 
-    while (x > 0.0) {
+    while (x > 0.0f) {
         printf("%d, x: %.20e\n", count, x);
-        x /= 2.0;
+        x /= 2.0f;
         count++;
     }
